libusbgx: fix bMaxPower handling
libusbgx expects the maximum power consumption value to be specified in
2mA units, but the kernel takes the "real" mA value and does the
conversion to 2mA units when constructing the High Speed descriptors.

Convert the value to and from the expected format when reading from or
writing to configfs.

This is sufficient for High Speed devices where the maximum possible
current draw is 500mA, but still does not permit SuperSpeed or faster
devices advertising current draw up to 900mA.